requirements.txt
asgiref==3.8.1
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.29.0
pillow==10.4.0
psycopg2==2.9.9
pymongo==4.8.0
python-dotenv==1.0.1
sqlparse==0.2.4
tzdata==2024.1
